+++ src/model/track.rs
@@ -342,3 +342,9 @@ impl ListItem for Track {
         Box::new(self.clone())
     }
 }
+
+impl PartialEq for Track {
+    fn eq(&self, other: &Self) -> bool {
+        self.id == other.id
+    }
+}