+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Special color for unavailable items
 - changelog with all the relevant user-facing changes to the project
+- Media control keys support for Windows and macOS
 
 ### Changed